+++ npm/bin/jscpd-rs.js
@@ -1,0 +1,6 @@
+#!/usr/bin/env node
+"use strict";
+
+const { runBinary } = require("../lib/run-binary");
+
+runBinary("jscpd", process.argv.slice(2));